replace student page with new page
